Feat : UserDao, UserRepository, UserViewModel 생성
Room에서 User를 위한 세팅
